moves test screen to new ui/test folder
