+++ lib/services/routing_service.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:lingo_jam/ui/favourites/favourites_screen.dart';
 import 'package:lingo_jam/ui/stations/stations_screen.dart';
-import 'package:lingo_jam/ui/screens/test_screen.dart';
+import 'package:lingo_jam/ui/test/test_screen.dart';
 
 class RoutingService {
   static const String home = '/';
